AI-HRMS/chatbot-api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
uuid==1.30
google-genai==0.2.2
